Fix some more prototypes (size_t != necessarily unsigned int).

+++ Library/libs/bzero.c
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <sys/types.h>
 
-void bzero(void *dest, int len)
+void bzero(void *dest, size_t len)
 {
    (void) memset(dest, '\0', len);
 }
